feat: add server side data fetching
