+++ src/lib/restClient/restClient.ts
@@ -1,26 +1,6 @@
-import type { Header } from '@/components/restClient/types';
+'use server';
 
-export type HttpMethod = 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE' | 'HEAD' | 'OPTIONS';
-
-interface RestClientParams<TBody> {
-  method: HttpMethod;
-  url: string;
-  body?: TBody;
-  headers?: Record<string, string> | Header[];
-}
-
-export interface RestClientResponse<TResponse> {
-  data: TResponse;
-  status: number;
-  statusText: string;
-  duration: number;
-}
-
-export interface RestClientError extends Error {
-  status?: number | null;
-  statusText?: string;
-  duration?: number | null;
-}
+import type { Header, RestClientParams, RestClientResponse } from '@/components/restClient/types';
 
 function normalizeHeaders(headers?: Record<string, string> | Header[]): Record<string, string> {
   if (!headers) return {};
@@ -40,9 +20,7 @@ export async function restClient<TResponse, TBody>({
   url,
   body,
   headers,
-}: RestClientParams<TBody>): Promise<RestClientResponse<TResponse>> {
-  const startTime = Date.now();
-
+}: RestClientParams<TBody>): Promise<RestClientResponse<TResponse> & { error?: string }> {
   const normalizedHeaders = normalizeHeaders(headers);
   const options: RequestInit = {
     method,
@@ -57,29 +35,43 @@ export async function restClient<TResponse, TBody>({
     }
   }
 
-  const response = await fetch(url, options);
-  const duration = Date.now() - startTime;
+  const startTime = Date.now();
 
-  if (!response.ok) {
-    const error = Object.assign(new Error(`Request failed with status ${response.status}`), {
+  try {
+    const response = await fetch(url, options);
+    const duration = Date.now() - startTime;
+
+    let data: TResponse;
+    try {
+      data = (await response.json()) as TResponse;
+    } catch {
+      data = (await response.text()) as unknown as TResponse;
+    }
+
+    if (!response.ok) {
+      return {
+        data,
+        status: response.status,
+        statusText: response.statusText,
+        duration,
+        error: `Request failed with status ${response.status}`,
+      };
+    }
+
+    return {
+      data,
       status: response.status,
       statusText: response.statusText,
       duration,
-    });
-    throw error;
+    };
+  } catch (err: unknown) {
+    const error = err as Error;
+    return {
+      data: null as unknown as TResponse,
+      status: 0,
+      statusText: 'Network error',
+      duration: 0,
+      error: error.message,
+    };
   }
-
-  let data: TResponse;
-  try {
-    data = (await response.json()) as TResponse;
-  } catch {
-    data = (await response.text()) as TResponse;
-  }
-
-  return {
-    data,
-    status: response.status,
-    statusText: response.statusText,
-    duration,
-  };
 }
